+++ Example/HelloSDK/HelloVC.swift
@@ -2,6 +2,12 @@ import UIKit
 import CoreLocation
 import MapGL
 
+/// contact us mapgl@2gis.com if you need one
+enum Constants {
+	static let apiKey = ""
+	static let directionsApiKey = ""
+}
+
 class HelloVC: UIViewController {
 
 	private static let cardViewHeight: CGFloat = 100
@@ -102,10 +108,9 @@ class HelloVC: UIViewController {
 			self.cardView.onRemove = nil
 			self.hideCardView()
 		}
-		let apiKey = "apiKey"
-		assert(apiKey != "apiKey", "contact us mapgl@2gis.com if you need one")
+		assert(!Constants.apiKey.isEmpty, "contact us mapgl@2gis.com if you need one")
 		self.map.show(
-			apiKey: apiKey,
+			apiKey: Constants.apiKey,
 			center: CLLocationCoordinate2D(latitude: 25.23584, longitude: 55.31878),
 			zoom: 16
 		) { error in
@@ -181,6 +186,18 @@ class HelloVC: UIViewController {
 		alert.popoverPresentationController?.sourceView = self.menuButton
 		alert.popoverPresentationController?.sourceRect = CGRect(x: self.menuButton.bounds.midX, y: -16, width: 0, height: 0)
 
+		let showRoute = UIAlertAction(title: "Show Route, hide in 10 sec", style: .default) { _ in
+			assert(!Constants.directionsApiKey.isEmpty, "contact us mapgl@2gis.com if you need one")
+			let directions = self.map.makeDirections(with: Constants.directionsApiKey)
+			directions.showCarRoute(points: [
+				self.map.mapCenter,
+				CLLocationCoordinate2D(latitude: 25.20, longitude: 55.4878),
+				CLLocationCoordinate2D(latitude: 25.20, longitude: 55.5278),
+			])
+			DispatchQueue.main.asyncAfter(deadline: .now() + 10) {
+				directions.clear()
+			}
+		}
 		let showMarkerAction = UIAlertAction(title: "Add Marker", style: .default) { _ in
 			let marker = Marker(
 				coordinates: self.map.mapCenter,
@@ -263,6 +280,7 @@ class HelloVC: UIViewController {
 
 		let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
 
+		alert.addAction(showRoute)
 		alert.addAction(showBuilding)
 		alert.addAction(showLabel)
 		alert.addAction(showCluster)